put pred flag uint8_t

+++ libavcodec/hevc.h
@@ -1017,7 +1017,7 @@ typedef struct MvField {
     Mv mv[2];
 #ifdef TEST_MV_POC
     int32_t poc[2];
-    uint32_t pred_flag;
+    uint8_t pred_flag;
 #else
     uint8_t pred_flag;
 #endif
